Fix exercise fret range selection; Styling changes,
The fix was to not call
`show_fret_range_modal.set(false);` on the onclick of the apply button. We did this after updating the exercise which cause the entire fret selection to be rerendered. I think then we call set on the disposed signal. Even though i got the error in the place where the signal was read in the when of the Show